+++ test/memex_web/live/pipeline_live_test.exs
@@ -17,7 +17,7 @@ defmodule MemexWeb.PipelineLiveTest do
   }
   @invalid_attrs %{
     description: nil,
-    tags_string: "invalid tags",
+    tags_string: "invalid_tag or_tags",
     slug: nil,
     visibility: nil
   }
@@ -130,7 +130,7 @@ defmodule MemexWeb.PipelineLiveTest do
         |> render_change(pipeline: @invalid_attrs)
 
       assert html =~ "can&#39;t be blank"
-      assert html =~ "tags must be comma-delimited"
+      assert html =~ "tags must be comma or space delimited"
 
       {:ok, _live, html} =
         show_live